+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Introduce a unified configuration abstraction
+- Update frequency and batch size configuration parameters
+- Improve sending assets log
 
 ### Fixed